fix(authtypes): treat kind as authoritative for auth domain config
The declared kind decides the provider; spec fields belonging to another
kind are ignored rather than rejected. Reads already decode leniently, so
this keeps the request path consistent with the stored-document path.
