+++ pkg/types/authtypes/domain.go
@@ -1,7 +1,6 @@
 package authtypes
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"regexp"
@@ -10,7 +9,6 @@ import (
 	"github.com/SigNoz/signoz/pkg/errors"
 	"github.com/SigNoz/signoz/pkg/types"
 	"github.com/SigNoz/signoz/pkg/valuer"
-	"github.com/swaggest/jsonschema-go"
 	"github.com/uptrace/bun"
 )
 
@@ -28,76 +26,6 @@ var (
 	ErrCodeAuthDomainMismatch      = errors.MustNewCode("auth_domain_mismatch")
 	ErrCodeAuthDomainNotFound      = errors.MustNewCode("auth_domain_not_found")
 	ErrCodeAuthDomainAlreadyExists = errors.MustNewCode("auth_domain_already_exists")
-)
-
-// authDomainConfigVariants is the single registry of authn provider kinds:
-// UnmarshalJSON, JSONSchemaOneOf and the discriminator mapping all derive from
-// it, so a new provider is one entry here plus its authn registration.
-//
-// rejectUnknownSpecFields decodes into a method-less alias of the spec: a type
-// carrying its own UnmarshalJSON consumes the bytes itself, which would bypass
-// DisallowUnknownFields. It only sees the spec's own fields, not those of
-// nested objects that unmarshal themselves.
-var authDomainConfigVariants = []authDomainConfigVariant{
-	{
-		kind: AuthNProviderSAML,
-		decodeSpec: func(data []byte) (any, error) {
-			spec := SamlConfig{}
-			if err := json.Unmarshal(data, &spec); err != nil {
-				return nil, err
-			}
-			return spec, nil
-		},
-		rejectUnknownSpecFields: func(data []byte) error {
-			type alias SamlConfig
-			decoder := json.NewDecoder(bytes.NewReader(data))
-			decoder.DisallowUnknownFields()
-			return decoder.Decode(new(alias))
-		},
-		schema:    authDomainConfigSAML{},
-		schemaRef: "#/components/schemas/AuthtypesAuthDomainConfigSAML",
-	},
-	{
-		kind: AuthNProviderGoogle,
-		decodeSpec: func(data []byte) (any, error) {
-			spec := GoogleConfig{}
-			if err := json.Unmarshal(data, &spec); err != nil {
-				return nil, err
-			}
-			return spec, nil
-		},
-		rejectUnknownSpecFields: func(data []byte) error {
-			type alias GoogleConfig
-			decoder := json.NewDecoder(bytes.NewReader(data))
-			decoder.DisallowUnknownFields()
-			return decoder.Decode(new(alias))
-		},
-		schema:    authDomainConfigGoogle{},
-		schemaRef: "#/components/schemas/AuthtypesAuthDomainConfigGoogle",
-	},
-	{
-		kind: AuthNProviderOIDC,
-		decodeSpec: func(data []byte) (any, error) {
-			spec := OIDCConfig{}
-			if err := json.Unmarshal(data, &spec); err != nil {
-				return nil, err
-			}
-			return spec, nil
-		},
-		rejectUnknownSpecFields: func(data []byte) error {
-			type alias OIDCConfig
-			decoder := json.NewDecoder(bytes.NewReader(data))
-			decoder.DisallowUnknownFields()
-			return decoder.Decode(new(alias))
-		},
-		schema:    authDomainConfigOIDC{},
-		schemaRef: "#/components/schemas/AuthtypesAuthDomainConfigOIDC",
-	},
-}
-
-var (
-	_ jsonschema.OneOfExposer = AuthDomainConfig{}
-	_ jsonschema.Preparer     = AuthDomainConfig{}
 )
 
 type GettableAuthDomain struct {
@@ -135,77 +63,9 @@ type StorableAuthDomain struct {
 	types.TimeAuditable
 }
 
-// StorableAuthDomainConfig is the JSON document persisted in StorableAuthDomain.Data.
-type StorableAuthDomainConfig struct {
-	Enabled     bool             `json:"enabled"`
-	Config      AuthDomainConfig `json:"config"`
-	RoleMapping *RoleMapping     `json:"roleMapping"`
-}
-
-type AuthDomainConfig struct {
-	Kind AuthNProvider `json:"kind" required:"true"`
-	Spec any           `json:"spec" required:"true"`
-
-	// rawSpec retains the undecoded spec bytes from UnmarshalJSON so the
-	// request path can run its foreign-field check without re-parsing the body.
-	rawSpec json.RawMessage
-}
-
-// authDomainConfigSAML is the OpenAPI schema for an AuthDomainConfig with kind=saml.
-type authDomainConfigSAML struct {
-	Kind AuthNProvider `json:"kind" description:"The kind of authn provider." required:"true"`
-	Spec SamlConfig    `json:"spec" description:"The saml configuration." required:"true"`
-}
-
-// authDomainConfigGoogle is the OpenAPI schema for an AuthDomainConfig with kind=google.
-type authDomainConfigGoogle struct {
-	Kind AuthNProvider `json:"kind" description:"The kind of authn provider." required:"true"`
-	Spec GoogleConfig  `json:"spec" description:"The google auth configuration." required:"true"`
-}
-
-// authDomainConfigOIDC is the OpenAPI schema for an AuthDomainConfig with kind=oidc.
-type authDomainConfigOIDC struct {
-	Kind AuthNProvider `json:"kind" description:"The kind of authn provider." required:"true"`
-	Spec OIDCConfig    `json:"spec" description:"The oidc configuration." required:"true"`
-}
-
-type authDomainConfigVariant struct {
-	kind                    AuthNProvider
-	decodeSpec              func(data []byte) (any, error)
-	rejectUnknownSpecFields func(data []byte) error
-	schema                  any
-	schemaRef               string
-}
-
 type AuthDomain struct {
 	storableAuthDomain       *StorableAuthDomain
 	storableAuthDomainConfig *StorableAuthDomainConfig
-}
-
-type AuthDomainStore interface {
-	// Get by id.
-	Get(context.Context, valuer.UUID) (*AuthDomain, error)
-
-	// Get by orgID and id.
-	GetByOrgIDAndID(context.Context, valuer.UUID, valuer.UUID) (*AuthDomain, error)
-
-	// Get by name.
-	GetByName(context.Context, string) (*AuthDomain, error)
-
-	// Get by name and orgID.
-	GetByNameAndOrgID(context.Context, string, valuer.UUID) (*AuthDomain, error)
-
-	// List org domains by orgID.
-	ListByOrgID(context.Context, valuer.UUID) ([]*AuthDomain, error)
-
-	// Create auth domain.
-	Create(context.Context, *AuthDomain) error
-
-	// Update by orgID and id.
-	Update(context.Context, *AuthDomain) error
-
-	// Delete by orgID and id.
-	Delete(context.Context, valuer.UUID, valuer.UUID) error
 }
 
 func NewAuthDomainFromPostableAuthDomain(postableAuthDomain *PostableAuthDomain, orgID valuer.UUID) (*AuthDomain, error) {
@@ -259,105 +119,6 @@ func NewGettableAuthDomainFromAuthDomain(authDomain *AuthDomain, authNProviderIn
 	}
 }
 
-// JSONSchemaOneOf returns the oneOf variants for the AuthDomainConfig discriminated union.
-// Each variant represents a different authn provider kind with its corresponding spec schema.
-func (AuthDomainConfig) JSONSchemaOneOf() []any {
-	oneOf := make([]any, len(authDomainConfigVariants))
-	for i, variant := range authDomainConfigVariants {
-		oneOf[i] = variant.schema
-	}
-
-	return oneOf
-}
-
-// PrepareJSONSchema marks the schema with x-signoz-discriminator;
-// signoz.attachDiscriminators promotes it to a real OpenAPI 3
-// discriminator after reflection.
-func (AuthDomainConfig) PrepareJSONSchema(schema *jsonschema.Schema) error {
-	if schema.ExtraProperties == nil {
-		schema.ExtraProperties = map[string]any{}
-	}
-
-	mapping := make(map[string]string, len(authDomainConfigVariants))
-	for _, variant := range authDomainConfigVariants {
-		mapping[variant.kind.StringValue()] = variant.schemaRef
-	}
-
-	schema.ExtraProperties["x-signoz-discriminator"] = map[string]any{
-		"propertyName": "kind",
-		"mapping":      mapping,
-	}
-
-	return nil
-}
-
-func (typ *AuthDomainConfig) UnmarshalJSON(data []byte) error {
-	var raw map[string]json.RawMessage
-	if err := json.Unmarshal(data, &raw); err != nil {
-		return errors.Wrapf(err, errors.TypeInvalidInput, ErrCodeAuthDomainInvalidConfig, "failed to unmarshal auth domain config")
-	}
-
-	kindData, ok := raw["kind"]
-	if !ok {
-		return errors.Newf(errors.TypeInvalidInput, ErrCodeAuthDomainInvalidConfig, "kind is required")
-	}
-
-	var kind AuthNProvider
-	if err := json.Unmarshal(kindData, &kind); err != nil {
-		return errors.Wrapf(err, errors.TypeInvalidInput, ErrCodeAuthDomainInvalidConfig, "failed to unmarshal kind")
-	}
-
-	specData, ok := raw["spec"]
-	if !ok {
-		return errors.Newf(errors.TypeInvalidInput, ErrCodeAuthDomainInvalidConfig, "spec is required")
-	}
-
-	for _, variant := range authDomainConfigVariants {
-		if variant.kind != kind {
-			continue
-		}
-
-		spec, err := variant.decodeSpec(specData)
-		if err != nil {
-			return err
-		}
-
-		typ.Kind = kind
-		typ.Spec = spec
-		typ.rawSpec = specData
-		return nil
-	}
-
-	return errors.Newf(errors.TypeInvalidInput, ErrCodeAuthDomainInvalidConfig, "invalid authn provider %q", kind.StringValue())
-}
-
-func (config AuthDomainConfig) SamlConfig() (SamlConfig, error) {
-	spec, ok := config.Spec.(SamlConfig)
-	if !ok {
-		return SamlConfig{}, errors.Newf(errors.TypeInternal, ErrCodeAuthDomainMismatch, "auth domain config is not saml")
-	}
-
-	return spec, nil
-}
-
-func (config AuthDomainConfig) GoogleConfig() (GoogleConfig, error) {
-	spec, ok := config.Spec.(GoogleConfig)
-	if !ok {
-		return GoogleConfig{}, errors.Newf(errors.TypeInternal, ErrCodeAuthDomainMismatch, "auth domain config is not google")
-	}
-
-	return spec, nil
-}
-
-func (config AuthDomainConfig) OIDCConfig() (OIDCConfig, error) {
-	spec, ok := config.Spec.(OIDCConfig)
-	if !ok {
-		return OIDCConfig{}, errors.Newf(errors.TypeInternal, ErrCodeAuthDomainMismatch, "auth domain config is not oidc")
-	}
-
-	return spec, nil
-}
-
 func (typ *AuthDomain) StorableAuthDomain() *StorableAuthDomain {
 	return typ.storableAuthDomain
 }
@@ -396,29 +157,6 @@ func (typ *AuthDomain) Update(updatableAuthDomain *UpdatableAuthDomain) error {
 	return nil
 }
 
-// rejectForeignSpecFields rejects a request whose config.spec carries fields the
-// declared kind does not define. Without it a spec belonging to another provider
-// decodes as a partial config of the declared kind whenever their required
-// fields overlap — an oidc spec sent as kind google loses its issuer and points
-// the domain at Google instead of the intended provider.
-//
-// This is a request-shape check by design: stored documents keep decoding
-// leniently so a rollback can still read a document a newer binary wrote, and so
-// removing a field later does not need a migration.
-func rejectForeignSpecFields(config AuthDomainConfig) error {
-	for _, variant := range authDomainConfigVariants {
-		if variant.kind != config.Kind {
-			continue
-		}
-
-		if err := variant.rejectUnknownSpecFields(config.rawSpec); err != nil {
-			return errors.Wrapf(err, errors.TypeInvalidInput, ErrCodeAuthDomainInvalidConfig, "invalid %q spec", config.Kind.StringValue())
-		}
-	}
-
-	return nil
-}
-
 func (typ *PostableAuthDomain) UnmarshalJSON(data []byte) error {
 	type Alias PostableAuthDomain
 
@@ -435,10 +173,6 @@ func (typ *PostableAuthDomain) UnmarshalJSON(data []byte) error {
 	// anything else), so a zero kind means the key was absent.
 	if temp.Config.Kind.IsZero() {
 		return errors.Newf(errors.TypeInvalidInput, ErrCodeAuthDomainInvalidConfig, "config is required")
-	}
-
-	if err := rejectForeignSpecFields(temp.Config); err != nil {
-		return err
 	}
 
 	*typ = PostableAuthDomain(temp)
@@ -459,10 +193,32 @@ func (typ *UpdatableAuthDomain) UnmarshalJSON(data []byte) error {
 		return errors.Newf(errors.TypeInvalidInput, ErrCodeAuthDomainInvalidConfig, "config is required")
 	}
 
-	if err := rejectForeignSpecFields(temp.Config); err != nil {
-		return err
-	}
-
 	*typ = UpdatableAuthDomain(temp)
 	return nil
+}
+
+type AuthDomainStore interface {
+	// Get by id.
+	Get(context.Context, valuer.UUID) (*AuthDomain, error)
+
+	// Get by orgID and id.
+	GetByOrgIDAndID(context.Context, valuer.UUID, valuer.UUID) (*AuthDomain, error)
+
+	// Get by name.
+	GetByName(context.Context, string) (*AuthDomain, error)
+
+	// Get by name and orgID.
+	GetByNameAndOrgID(context.Context, string, valuer.UUID) (*AuthDomain, error)
+
+	// List org domains by orgID.
+	ListByOrgID(context.Context, valuer.UUID) ([]*AuthDomain, error)
+
+	// Create auth domain.
+	Create(context.Context, *AuthDomain) error
+
+	// Update by orgID and id.
+	Update(context.Context, *AuthDomain) error
+
+	// Delete by orgID and id.
+	Delete(context.Context, valuer.UUID, valuer.UUID) error
 }
