+++ pkg/types/authtypes/domain_test.go
@@ -10,57 +10,44 @@ import (
 
 func TestPostableAuthDomainSpecFields(t *testing.T) {
 	testCases := []struct {
-		name          string
-		body          string
-		expectedError bool
+		name         string
+		body         string
+		expectedKind AuthNProvider
 	}{
 		{
-			name:          "SAMLSpec",
-			body:          `{"name":"a.test","enabled":true,"config":{"kind":"saml","spec":{"entityId":"e","location":"l","certificate":"c"}}}`,
-			expectedError: false,
+			name:         "SAMLSpec",
+			body:         `{"name":"a.test","enabled":true,"config":{"kind":"saml","spec":{"entityId":"e","location":"l","certificate":"c"}}}`,
+			expectedKind: AuthNProviderSAML,
 		},
 		{
-			name:          "SAMLSpecWithAttributeMapping",
-			body:          `{"name":"a.test","enabled":true,"config":{"kind":"saml","spec":{"entityId":"e","location":"l","certificate":"c","insecureSkipAuthNRequestsSigned":true,"attributeMapping":{"email":"mail"}}}}`,
-			expectedError: false,
+			name:         "SAMLSpecWithAttributeMapping",
+			body:         `{"name":"a.test","enabled":true,"config":{"kind":"saml","spec":{"entityId":"e","location":"l","certificate":"c","insecureSkipAuthNRequestsSigned":true,"attributeMapping":{"email":"mail"}}}}`,
+			expectedKind: AuthNProviderSAML,
 		},
 		{
-			name:          "GoogleSpec",
-			body:          `{"name":"a.test","enabled":true,"config":{"kind":"google","spec":{"clientId":"c","clientSecret":"s","insecureSkipEmailVerified":false}}}`,
-			expectedError: false,
+			name:         "GoogleSpec",
+			body:         `{"name":"a.test","enabled":true,"config":{"kind":"google","spec":{"clientId":"c","clientSecret":"s","insecureSkipEmailVerified":false}}}`,
+			expectedKind: AuthNProviderGoogle,
 		},
 		{
-			name:          "GoogleSpecWithWorkspaceGroups",
-			body:          `{"name":"a.test","enabled":true,"config":{"kind":"google","spec":{"clientId":"c","clientSecret":"s","fetchGroups":true,"serviceAccountJson":"{}","domainToAdminEmail":{"*":"admin@a.test"},"fetchTransitiveGroupMembership":true,"allowedGroups":["g@a.test"]}}}`,
-			expectedError: false,
+			name:         "GoogleSpecWithWorkspaceGroups",
+			body:         `{"name":"a.test","enabled":true,"config":{"kind":"google","spec":{"clientId":"c","clientSecret":"s","fetchGroups":true,"serviceAccountJson":"{}","domainToAdminEmail":{"*":"admin@a.test"},"fetchTransitiveGroupMembership":true,"allowedGroups":["g@a.test"]}}}`,
+			expectedKind: AuthNProviderGoogle,
 		},
 		{
-			name:          "OIDCSpec",
-			body:          `{"name":"a.test","enabled":true,"config":{"kind":"oidc","spec":{"issuer":"https://issuer.a.test","clientId":"c","clientSecret":"s"}}}`,
-			expectedError: false,
+			name:         "OIDCSpec",
+			body:         `{"name":"a.test","enabled":true,"config":{"kind":"oidc","spec":{"issuer":"https://issuer.a.test","clientId":"c","clientSecret":"s"}}}`,
+			expectedKind: AuthNProviderOIDC,
 		},
 		{
-			// clientId and clientSecret are google's only required fields, so an
-			// oidc spec satisfies them and would otherwise be accepted with its
-			// issuer silently dropped.
-			name:          "OIDCSpecUnderGoogleKind",
-			body:          `{"name":"a.test","enabled":true,"config":{"kind":"google","spec":{"issuer":"https://issuer.a.test","clientId":"c","clientSecret":"s","claimMapping":{"email":"mail"}}}}`,
-			expectedError: true,
+			name:         "OIDCFieldsUnderGoogleKindConfigureGoogle",
+			body:         `{"name":"a.test","enabled":true,"config":{"kind":"google","spec":{"issuer":"https://issuer.a.test","clientId":"c","clientSecret":"s","claimMapping":{"email":"mail"}}}}`,
+			expectedKind: AuthNProviderGoogle,
 		},
 		{
-			name:          "GoogleSpecUnderOIDCKind",
-			body:          `{"name":"a.test","enabled":true,"config":{"kind":"oidc","spec":{"issuer":"https://issuer.a.test","clientId":"c","clientSecret":"s","fetchGroups":true}}}`,
-			expectedError: true,
-		},
-		{
-			name:          "SAMLSpecUnderGoogleKind",
-			body:          `{"name":"a.test","enabled":true,"config":{"kind":"google","spec":{"clientId":"c","clientSecret":"s","entityId":"e"}}}`,
-			expectedError: true,
-		},
-		{
-			name:          "UnrecognizedFieldInGoogleSpec",
-			body:          `{"name":"a.test","enabled":true,"config":{"kind":"google","spec":{"clientId":"c","clientSecret":"s","redirectURI":"https://a.test/cb"}}}`,
-			expectedError: true,
+			name:         "GoogleOnlyFieldUnderGoogleKindIgnored",
+			body:         `{"name":"a.test","enabled":true,"config":{"kind":"google","spec":{"clientId":"c","clientSecret":"s","redirectURI":"https://a.test/cb"}}}`,
+			expectedKind: AuthNProviderGoogle,
 		},
 	}
 
@@ -68,31 +55,23 @@ func TestPostableAuthDomainSpecFields(t *testing.T) {
 		t.Run(testCase.name, func(t *testing.T) {
 			postable := new(PostableAuthDomain)
 			err := json.Unmarshal([]byte(testCase.body), postable)
+			require.NoError(t, err)
 
-			if testCase.expectedError {
-				assert.Error(t, err)
-				return
-			}
-
-			assert.NoError(t, err)
+			assert.Equal(t, testCase.expectedKind, postable.Config.Kind)
 		})
 	}
 }
 
-// The enforcement toggle and the edit modal both PUT a full replacement, so the
-// update body needs the same check as the create body.
+// The enforcement toggle and the edit modal both PUT a full replacement, so the update body decodes with the same kind-authoritative rules as the create body.
 func TestUpdatableAuthDomainSpecFields(t *testing.T) {
 	updatable := new(UpdatableAuthDomain)
 	err := json.Unmarshal([]byte(`{"enabled":true,"config":{"kind":"google","spec":{"issuer":"https://issuer.a.test","clientId":"c","clientSecret":"s"}}}`), updatable)
-	assert.Error(t, err)
+	require.NoError(t, err)
 
-	updatable = new(UpdatableAuthDomain)
-	err = json.Unmarshal([]byte(`{"enabled":true,"config":{"kind":"google","spec":{"clientId":"c","clientSecret":"s"}}}`), updatable)
-	assert.NoError(t, err)
+	assert.Equal(t, AuthNProviderGoogle, updatable.Config.Kind)
 }
 
-// Reads stay lenient on purpose: a document written by a newer binary has to keep
-// loading after a rollback, and dropping a field must not need a migration.
+// Reads stay lenient on purpose: a document written by a newer binary has to keep loading after a rollback, and dropping a field must not need a migration.
 func TestNewAuthDomainFromStorableAuthDomainKeepsUnknownSpecFields(t *testing.T) {
 	storable := &StorableAuthDomain{
 		Data: `{"enabled":true,"config":{"kind":"google","spec":{"clientId":"c","clientSecret":"s","fieldFromANewerVersion":"x"}}}`,
